[MOD] Hello dear joss for some reason my whole reposity was deleted in the github and had to clone it again.
I had to update the scene because the masses and frictioncoefficient values were reseted to back to what they were. But rn the green and yellow ball have the same friction but their friction vector's magnitude is changing based off of mass too. If you watch the video you'll see that they are the same. but in this solution they differ.
Is it right ?
but their gravity force is correct.